fix: minor changes
About pnpm format and lint i think it was mentioned earlier but i forgot Sorry about that. I changed the width to that suggested by you.
I also replaced functions used as a type in menu.
About the text center earlier when the width was different text "you haven't added any satelites" wasn't always centered correctly justify-center didn't seemed to work so i added text-center  so it would be centered properly i checked if removing justify-center would change anything and it didn't so i'm not sure if i should leave it as it is or remove one of them.
I used pnpm lint and there were no problems
pnpm format seemed to modify a lot of files but on github desktop those files had no changes  so i discarded them not sure if that's how it should work.